Add serialization around client interactions
In order to compare as fairly as we can against vanilla redis, we need
to ensure that we are not handling client in parallel. All client
interactions come in via readQueryFromClient() so this function now
takes a spin lock at its beginning and releases it on the way out. We
use a spin lock to avoid invoking the scheduler (like a mutex would) and
simply moving the scheduler latency to the mutex.

Signed-off-by: Eric B. Munson <munsoner@bu.edu>