Make event handler send responses before looking for new work
This was the cause of our 100ms latencies. Redis queues responses as it
processes its command buffer, but queueing a response does not imply
that the function to send them is called. This would normally cause our
implementation to fail, except that redis also maintains several
internal timers. One of these timers has a period of about 100ms and
part of what it does when this timer fires is to send any pending
responses to clients. This explains our ~100ms latencies using our
initial implementaion.

Signed-off-by: Eric B Munson <munsoner@bu.edu>